Import of release hp2xx-3.4.3_a08.tar.gz 3.4.3a8:
moved the vertex adding stuff to a new function that updates the extents of
the drawing as well - previously vertices that were only used as boundaries
for filling, not edging, polygon, did not enter into the size calculation.
Tried to achieve a more sensible initialization of AC - 0,0 as in alpha7 is
plain stupid, but the "physical extents" is a somewhat dubious thing in hp2xx,
and the actual limits are not yet known when the IN or plain AC is seen.
Modified the warning message code to make sure that the "unsupported charset"
message is printed only once per SD, not once per character.
The seeming IW brokenness was not in hp2xx but in the program that created the
strange test file, so i have reverted the change from a7, which makes some
correct files display again.
Some fixes by Bengt-Arne Fjellner: usage message was printing garbage
in description of -p (because of penwidth type change), EMF files were
upside down, and microsoft platforms seem to need a replacement for
rint/lrint (actually some non-MS platforms may need this as well - not sure
if everybody does obey C99 already ?).